initial commit
This code was a freshmen engineering course project. Original folder contained large mp4 files. For the videos, the YouTube links are included in "workscited.txt". using wget in a code like this would be better for portability, but was not implemented initially, so this code searches for the mp4 files in your path instead.